Bug 80758  - Default caching of static resources should be more than 30 days
Changed the default value for max-age to 2595600(30 days and 1 hour). Also fixed the case where Expires header would get set incorrectly if the expiresValue in msec exceeds the max Integer value.

http://bugzilla.zimbra.com/show_bug.cgi?id=80758

+++ ZimbraWebClient/src/com/zimbra/webClient/filters/SetHeaderFilter.java
@@ -497,7 +497,7 @@ public final class SetHeaderFilter extends com.zimbra.cs.servlet.SetHeaderFilter
         if (expiresValue > 0 && req.getMethod().equals("GET")) {
             TimeZone gmt = TimeZone.getTimeZone(GMT);
             long now = System.currentTimeMillis();
-            Date expiresDate = new Date(now + (expiresValue * 1000));
+            Date expiresDate = new Date(now + (expiresValue * (long)1000));
             DateFormat df = new SimpleDateFormat(TIME_FORMAT);
             df.setTimeZone(gmt);
             resp.setHeader(HEADER_EXPIRES, df.format(expiresDate) + " " + GMT);
